pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dotenv = "^0.17.0"
requests = "^2.24.0"
gunicorn = "^20.1.0"
pymongo = {version = "^3.11.4", extras = ["srv", "tls"]}
mongomock = "^3.22.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
selenium = "^3.3.0"
pylint = "^2.8.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"